Создан дополнительный компонент для навигпции
